fixup :review comments
