+++ robot-server/.flake8
@@ -34,4 +34,3 @@ per-file-ignores =
     tests/service/*:ANN,D
     tests/conftest.py:ANN,D
     tests/test_util.py:ANN,D
-    tests/integration/protocol_files.py:E501
